python/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fx_intraday_ai"
version = "0.1.0"
description = "Intraday FX multi-task AI pipeline for EURUSD/USDJPY/GBPUSD on 5-minute bars."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [
    {name = "Codex"},
]
dependencies = [
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.4.0",
    "torch>=2.2.0",
    "lightgbm>=4.1.0",
    "ta>=0.11.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.6.0",
    "mlflow>=2.9.0",
    "rich>=13.7.0",
    "requests>=2.31.0",
    "click>=8.1.3",
    "numba>=0.58.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
